+++ examples/baiducloud-acl/main.tf
@@ -1,4 +1,11 @@
-provider "baiducloud" {}
+provider "baiducloud" {
+  # option config, you can use assume role as the operation account
+  #assume_role {
+  #  account_id = "your-account-id"
+  #  role_name = "your-role-name"
+  #}
+}
+
 
 data "baiducloud_zones" "default" {}