+++ examples/baiducloud-cfc/main.tf
@@ -1,4 +1,10 @@
-provider "baiducloud" {}
+provider "baiducloud" {
+  # option config, you can use assume role as the operation account
+  #assume_role {
+  #  account_id = "your-account-id"
+  #  role_name = "your-role-name"
+  #}
+}
 
 resource "baiducloud_bos_bucket" "default" {
   bucket = "terraform-bucket-20191106"