+++ examples/baiducloud-rds/main.tf
@@ -1,4 +1,10 @@
-provider "baiducloud" {}
+provider "baiducloud" {
+  # option config, you can use assume role as the operation account
+  #assume_role {
+  #  account_id = "your-account-id"
+  #  role_name = "your-role-name"
+  #}
+}
 
 resource "baiducloud_rds_instance" "default" {
   billing = {